pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = "Дипломный проект: календарь"
authors = ["Tuman8 <i-a-l-e-x@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.3"

#[tool.poetry.group.dev.dependencies]
#pre-commit = "^2.20.0"
djangorestframework = "^3.13.1"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
py_version = 310
profile = 'django'
line_length = 120
multi_line_output = 2
known_first_party = ['todolist', 'core']

[tool.flake8]
max-line-length = 120
per-file-ignores = """
    */migrations/*: E501,
"""
